fix: API router fixes
