+++ src/webService/api/V2/users.js
@@ -1,21 +1,25 @@
 //------------------------------------------[ Imports ]------------------------------------------\\
 const express = require('express');
 const router = express.Router()
-const responder = require('./').responder
+const responder = require('./utils/responder')
 
 
 
-//------------------------------------[ Endpoints ]------------------------------------\\
+//-----------------------------------------[ Endpoints ]-----------------------------------------\\
 
+// GET/FETCH Users:
 router.get('/users/:userId', async (req, res) => {
     const fetchId = req.params.userId
     if(!fetchId) return responder.errored(res, `invalid "userId" provided`)
     else return responder.succeeded(res, `[GET] Request send for user ${fetchId}`)
 })
 
-
+// DELETE/REMOVE Users:
 router.delete('/user/:userId', async (req, res) => {
     const fetchId = req.params.userId
     if(!fetchId) return responder.errored(res, `invalid "userId" provided`)
     else return responder.succeeded(res, `[DELETE] Request send for user ${fetchId}`)
 })
+
+//-------------------------------------[ Export Endpoints ]-------------------------------------\\
+module.exports = router
