+++ src/webService/webService.js
@@ -39,7 +39,7 @@ const discordAuthRouter = require('./discordAuth.js');
 app.use('/api', discordAuthRouter); // or just app.use(discordAuthRouter);
 
 // Router - API V2:
-const apiV2 = require('./api/V2')
+const apiV2 = require('./api/V2/index.js')
 app.use('/api/v2', apiV2)
 
 // Initialize Port:
